fix: fixing rendering issues related to zustand
