+++ frontend/src/app/layout.tsx
@@ -1,12 +1,7 @@
 import type { Metadata } from 'next'
-import { Geist, Geist_Mono, Inter, Poppins } from 'next/font/google'
+import { Inter, Poppins } from 'next/font/google'
 import './globals.css'
 import { ContextProvider } from '../contextAPI/context'
-
-const geistSans = Geist({
-  variable: '--font-geist-sans',
-  subsets: ['latin'],
-})
 
 const inter = Inter({
   subsets: ['latin'],
@@ -21,11 +16,6 @@ const poppins = Poppins({
   display: 'swap',
 })
 
-const geistMono = Geist_Mono({
-  variable: '--font-geist-mono',
-  subsets: ['latin'],
-})
-
 export const metadata: Metadata = {
   title: 'Create Next App',
   description: 'Generated by create next app',
@@ -37,14 +27,10 @@ export default function RootLayout({
   children: React.ReactNode
 }>) {
   return (
-    <ContextProvider>
-      <html lang="en">
-        <body
-          className={`${geistSans.variable} ${geistMono.variable} ${inter.variable} ${poppins.variable} antialiased`}
-        >
-          {children}
-        </body>
-      </html>
-    </ContextProvider>
+    <html lang="en">
+      <body className={`${inter.variable} ${poppins.variable} antialiased`}>
+        <ContextProvider>{children}</ContextProvider>
+      </body>
+    </html>
   )
 }
